+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# 0.1.1 (Sep 4th 2025)
+# 0.1.1 (Sep 6th 2025)
 
 - Updated dependencies in package.json:
-  - @dwtechs/checkard to 3.5.0
+  - @dwtechs/checkard to 3.5.1
   - @dwtechs/hashitaka to 0.3.1
+- Improved error messages using Hashitaka and Checkard errors
 
 # 0.1.0 (Aug 10th 2025)